requirements.txt
discord.py==2.0.1
googletrans==3.0.0
